requirements.txt
Flask==3.0.3
gspread==6.1.4
oauth2client==4.1.3
google-api-python-client==2.141.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
requests==2.32.3
gunicorn==22.0.0
